+++ changelog.rst
@@ -27,7 +27,7 @@ ChangeLog
 - remove decorator module.
 - add foreigns attributes in Configurable which allows to add not specified parameters given by conf resources.
 - add autoconf attribute in Configurable, getconfigurables and applyconfiguration functions.
-- rename get_conf, set_conf, to_configure and apply_configuration to getconf, setconf, toconfigure and applyconfiguration.
+- rename get_conf, set_conf, to_configure and apply_configuration to getconf, setconf, targets and applyconfiguration.
 - add Configurable.safe attribute in order to execute configuration in an unsafe context if necessary.
 - add the configurable Logger useful to ease management of complex logging needs.
 
@@ -64,7 +64,7 @@ ChangeLog
 0.1.7 (2015/07/22)
 ------------------
 
-- fix bug about toconfigure parameter.
+- fix bug about targets parameter.
 - update README in fixing the example.
 
 0.1.6 (2015/06/13)